Add mostly working asset loading benchmark
Counter idea taken from https://github.com/darthdeus/comfy/pull/30

+++ generate-assets.sh
@@ -3,7 +3,11 @@ set -euxo pipefail
 
 dir="comfy-asset-benchmark/assets"
 
-for i in $(seq 1 100); do
+if [ -f "$dir/comfy-1000.png" ]; then
+  exit 0
+fi
+
+for i in $(seq 1 1000); do
   cp "$dir/_base-comfy.png" "$dir/comfy-$i.png"
   cp "$dir/_base-music.wav" "$dir/music-$i.wav"
 done
